Add message forwarding functionality to Kanboard
This update includes the creation of a new "ContextMenuForwardToKanboard" and a utility class "Kanboard", which manage the forwarding of messages to Kanboard. The "forwardMessageToKanboard" method has been extracted from "TagAdded" into "Kanboard" for better code organization. These changes also include some minor adjustments like changing the permission called in "AltitudeBot" and implementing the use of @Getter in the same class.